+++ x/leverage/types/flag.go
@@ -1,0 +1,6 @@
+package types
+
+const FlagEnableLiquidatorQuery = "enable-liquidator-query"
+
+// EnableLiquidator must be set to true to enable QueryLiquidationTargets
+var EnableLiquidator *bool